adjust to code style (move class to file, add braces)
